+++ Strange_Lovers/DrivenPart/TargetLightValueControl/PIDLightValCtrl.h
@@ -17,7 +17,7 @@ typedef struct{
 
 
 extern void PLVC_init(PIDLightValCtrl *self,PIDLightValCtrlParm *parm);
-extern int PLVC_calcCtrlVal(PIDLightValCtrl *self,float targLightVal,float LightVal,float time);
+extern float PLVC_calcCtrlVal(PIDLightValCtrl *self,float targLightVal,float LightVal,float time);
 extern void PLVC_setCtrlParm(PIDLightValCtrl *self,PIDLightValCtrlParm *parm);
 extern PIDLightValCtrlParm* PLVC_getCtrlParm(PIDLightValCtrl *self);
 extern void PLVC_reset(PIDLightValCtrl *self);
